refactor(auth): improve folder structure of SignIn
